fix(move): move Input/SearchBar to legacy
