+++ src/drawer/demos/Drawer.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Story, Meta } from '@storybook/react/types-6-0';
 import Button from '../../legacy/button';
-import Input from '../../components/input';
+import Input from '../../legacy/input';
 import { DrawerProps } from '../interfaces';
 import Drawer from '..';
 import '../style';
