+++ src/form/demos/Form.stories.tsx
@@ -5,7 +5,7 @@ import Item from '../Item';
 import Docs from './FormPage';
 import { Props, FormItemProps } from '../interface';
 import '../style';
-import '../../components/input/style';
+import '../../legacy/input/style';
 import '../../legacy/button/style';
 import '../style/demo.stories.less';
 import { Button, Input, Modal } from '../../index';
