+++ src/legacy/search-bar/interfaces.ts
