+++ src/legacy/search-bar/locales/en-US.ts
