+++ src/legacy/search-bar/locales/zh-CN.ts
