+++ src/legacy/search-bar/style/index.less
