+++ src/legacy/search-bar/style/index.ts
