+++ src/locales/zh-CN.ts
@@ -8,7 +8,7 @@ import listPickerLocale from '../list-picker/locales/zh-CN';
 import modalLocale from '../modal/locales/zh-CN';
 import timePickerLocale from '../time-picker/locales/zh-CN';
 import timeSelectorLocale from '../time-selector/locales/zh-CN';
-import searchBarLocale from '../search-bar/locales/zh-CN';
+import searchBarLocale from '../legacy/search-bar/locales/zh-CN';
 import cascaderLocale from '../cascader/locales/zh-CN';
 import paginationLocale from '../pagination/locales/zh-CN';
 import tableLocale from '../table/locales/zh-CN';
