+++ src/pagination/style/index.ts
@@ -1,3 +1,3 @@
 import './index.less';
-import '../../components/input/style/index';
+import '../../legacy/input/style/index';
 import '../../components/select/style/index';
