+++ src/table/style/index.ts
@@ -4,5 +4,5 @@ import '../../checkbox/style/index';
 import '../../pagination/style/index';
 import '../../tooltip/style/index';
 import '../../popover/style/index';
-import '../../search-bar/style/index';
+import '../../legacy/search-bar/style/index';
 import '../../components/empty/style/index';
